Avoid output if custom chooser not set
If a custom type "photo" of type chooser is created, but it is not set, and template includes `{_photo before="" after="" size="thumbnail"}` WP function ` wp_get_attachment_link` returns "Attachment missing", which is displayed by MCM. I expect no output, if a choser field is not set. This is what this PR does.